[BACKEND] Changing DB structure
Previously each team contained an array of its users. Instead, we create hardware and team in isolation, and
each user has a single reference to hardware/team. Makes it a bit cleaner since the DB is now normalized, and
we can now validate the user <-> team reference (not possible to do with an array of IDs).

Also changed the retirement/unretriement workflow. Instead we simply delete a user, and their points are then
retained as 'retired' points for their team (until the reset). To 'unretire' a user, simply recreate them. Since
the retired points are no longer referencing a user, this is a bit simpler to manage.

Frontend should now also be unblocked, since teams no longer need to find a list of users, and instead we can
populate the user with a dropdown menu of teams (as we do for the hardware).