[BACKEND] Reducing the duplicate REST requests for stats retrieval
Due to some server-side caching by the Folding@Home API, we have been sending two requests every update to get
new stats for each user. While I haven't been able to figure out how to get a non-cached response each time
(the 'Cache-Control: no-cache' header does not work), I've added a local cache which compares the previous
response to the current one. If a change is found, we don't bother sending a second request.

It is not a perfect solution, but at least reduces the number of requests.