[BUG] Not handling LARS DB incorrectly having duplicate GPU models
The assumption is that LARS should have a single entry per unique GPU model,
but it can sometimes be split into multiple. We enforced unique keys, so gathering
the data from LARS was causing a failure.

Now we only keep the first instance of an entry. If more are found, they are ignored.
This should be acceptable since the first entry will have the highest PPD for the GPU.