[BUG] Rate-limitting hits when resetting stats
Since we have the last update of stats at xx:55, then attempt to get latest stats again at xx:57 for the reset
we get rate-limited by the Stanford Folding@Home API. Instead, we'll just use the stats from xx:55 for the reset.

Should be fine since we do another reset when the competition starts again on the 3rd of the month.